+++ crates/eww/src/widgets/widget_definitions.rs
@@ -1200,6 +1200,7 @@ fn build_systray(bargs: &mut BuilderArgs) -> Result<gtk::Box> {
     let props = Rc::new(systray::Props::new());
     let props_clone = props.clone(); // copies for def_widget
     let props_clone2 = props.clone(); // copies for def_widget
+    let props_clone3 = props.clone(); // copies for def_widget
 
     def_widget!(bargs, _g, gtk_widget, {
         // @prop spacing - spacing between elements
@@ -1219,6 +1220,10 @@ fn build_systray(bargs: &mut BuilderArgs) -> Result<gtk::Box> {
         // @prop prepend-new - prepend new icons.
         prop(prepend_new: as_bool = true) {
             *props_clone2.prepend_new.borrow_mut() = prepend_new;
+        },
+        // @prop visible-empty - visibility of the widget when the systray is empty
+        prop(visible_empty: as_bool = true) {
+            *props_clone3.visible_empty.borrow_mut() = visible_empty;
         },
     });