+++ CHANGELOG.md
@@ -6,10 +6,12 @@ All notable changes to eww will be listed here, starting at changes since versio
 
 ### Fixes
 - Fix and refactor nix flake (By: w-lfchen)
+- Fix deleting items from tray (By: vnva)
 
 ### Features
 - Add `:truncate` property to labels, disabled by default (except in cases where truncation would be enabled in version `0.5.0` and before) (By: Rayzeq).
 - Add support for `:hover` css selectors for tray items (By: zeapoz)
+- Add `:visible-empty` property for `systray` widget (By: vnva)
 
 ## [0.6.0] (21.04.2024)